fix: change cookies to be saved even in a non-secure domain
Previously, if the system was exposed in a non-secure domain,
log-in and sign-in would fail as their cookies would be
discarded after the operation.

Now, it is possible to host the system in a machine with a
non-secure domain, exposing the service to a local network by
using the ip of the host machine in that network as the hostname.